docs: improve onboarding and game design skills based on dogfooding
