+++ .agents/skills/gbl-backend-computations/SKILL.md
@@ -31,9 +31,23 @@ Every hook: `(facts, payload) => OutputFacts`. Payloads carry what you need (`ga
 - Side channels in any hook's return: `events` (drive achievements), `notifications`/`globalNotification` (client toasts), `actions` (extra audit rows), `updatedPeriodFacts`/`updatedSegmentFacts` (from the action reducer).
 - Escape hatch only when facts blobs are not enough: optional `updateDBAfterInitialize` / `updateDBBeforeActivation` / `updateDBAfterEnd` / `updateDBAfterApply` receive the open Prisma transaction (use with tables you added in `prisma/schema/specific.prisma`).
 
+> [!WARNING]
+> **Never use `@prisma/client` enums in code that reaches the frontend.** Next.js stubs backend-only imports for the client bundle, so `DB.GameStatus.RESULTS` evaluates to `undefined` at runtime and crashes the page. In shared utilities imported by both server and client (like `packages/platform/src/lib/util.ts`), use string literals (`'RESULTS'`, `'PAUSED'`, etc.) or the GraphQL-generated enum from `src/graphql/generated/ops.ts`. Keep Prisma imports as `import type` when the file is consumed by frontend code.
+
 ## Facts types + validation
 
 Define types and yup schemas for `GameFacts`, `PeriodFacts`, `PeriodSegmentFacts`, `PlayerFacts` in `src/types/` (copy the demo game's file layout). The schemas gate admin inputs at the API boundary — the DB accepts any JSON, so schemas are the only validation.
+
+> [!TIP]
+> **Create a safe `parseFacts<T>()` helper early.** The `facts` column is `JsonValue` — it may be `null`, a raw object, or a double-stringified JSON string. Wrap every facts read in a helper like:
+> ```ts
+> function parseFacts<T>(raw: unknown, fallback: T): T {
+>   if (!raw) return fallback;
+>   try { return typeof raw === 'string' ? JSON.parse(raw) : raw as T; }
+>   catch { return fallback; }
+> }
+> ```
+> Use it in admin reports, cockpit views, and result services. Without it, uninitialised periods/segments will crash the UI with `JSON.parse(null)`.
 
 ## Seed
