+++ packages/platform/src/lib/util.ts
@@ -50,60 +50,6 @@ export function withPercentChange(value: number, percentChange: number) {
   return value * (1 + percentChange)
 }
 
-export enum STATUS {
-  ACTIVE = 'ACTIVE',
-  PAUSED = 'PAUSED',
-  SCHEDULED = 'SCHEDULED',
-  COMPLETED = 'COMPLETED',
-  RESULTS = 'RESULTS',
-}
-
-export function computePeriodStatus(
-  game: DB.Game,
-  periodIndex: number
-): string {
-  if (
-    typeof game.activePeriodIx === 'number' &&
-    game.status === DB.GameStatus.RESULTS
-      ? game.activePeriodIx - 1 === periodIndex
-      : game.activePeriodIx === periodIndex
-  ) {
-    if (game.status === DB.GameStatus.PAUSED) return STATUS.PAUSED
-    if (game.status === DB.GameStatus.RESULTS) return STATUS.RESULTS
-    return STATUS.ACTIVE
-  }
-
-  if (
-    typeof game.activePeriodIx === 'number' &&
-    game.activePeriodIx <= periodIndex
-  )
-    return STATUS.SCHEDULED
-
-  return STATUS.COMPLETED
-}
-
-export function computeSegmentStatus(
-  game: DB.Game,
-  period: DB.Period,
-  segmentIndex: number
-): string {
-  if (
-    ![
-      DB.GameStatus.PAUSED,
-      DB.GameStatus.PREPARATION,
-      DB.GameStatus.CONSOLIDATION,
-      DB.GameStatus.RESULTS,
-    ].includes(game.status as any) &&
-    period.activeSegmentIx === segmentIndex
-  )
-    return STATUS.ACTIVE
-
-  if (period.activeSegmentIx > -2 && period.activeSegmentIx < segmentIndex)
-    return STATUS.SCHEDULED
-
-  return STATUS.COMPLETED
-}
-
 export async function withRetry<T>(
   fn: () => Promise<T>,
   retries = 3
