+++ packages/ui/src/index.ts
@@ -11,3 +11,4 @@ export * from '~/components/Timeline'
 export * from '~/components/TimelineAdmin'
 export * from '~/components/TradingForm'
 export * from '~/components/XpBar'
+export * from '~/lib/status'
